# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/fetch-app-json:
    post:
      operationId: fetchAppJson
      tags: [admin]
      summary: Fetch and parse app.json from a GitHub repo URL (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchAppJsonInput"
      responses:
        "200":
          description: Parsed app.json data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppJsonResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/deployments/{id}/suspend:
    post:
      operationId: suspendDeployment
      tags: [admin]
      summary: Suspend a deployment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
...
    Deployment:
      type: object
      required: [id, userId, templateId, templateName, botName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: integer
        templateName:
          type: string
        botName:
          type: string
        herokuAppId:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, building, online, offline, error, suspended]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    FetchAppJsonInput:
      type: object
      required: [repoUrl]
      properties:
        repoUrl:
          type: string
    AppJsonResult:
      type: object
      required: [name, raw]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        logo:
          type: ["string", "null"]
        keywords:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: object
        raw:
          type: object
...